refactor(entrypoint): move execution logic into Lotto.run()
integrate main game loop inside Lotto companion object for cleaner structure and improved readability"
